KSENROLL-12051: created registration engine unit test. This tests a successful adding of a course through the reg engine.
git-svn-id: https://svn.kuali.org/repos/student/enrollment/ks-enroll/trunk@68236 0ac02d80-fb3f-11dc-a574-c57244a850eb

+++ ks-enroll-impl/src/main/java/org/kuali/student/enrollment/registration/search/service/impl/CourseRegistrationSearchServiceImpl.java
@@ -664,14 +664,14 @@ public class CourseRegistrationSearchServiceImpl extends SearchServiceAbstractHa
             SearchResultRowInfo row = new SearchResultRowInfo();
             row.addCell(SearchResultColumns.AO_ID, (String) resultRow[i++]);
             row.addCell(SearchResultColumns.AO_TYPE, (String) resultRow[i++]);
-            row.addCell(SearchResultColumns.AO_MAX_SEATS, resultRow[i] == null ? null : ((BigDecimal) resultRow[i]).toString());
+            row.addCell(SearchResultColumns.AO_MAX_SEATS, resultRow[i] == null ? null : resultRow[i].toString());
             i++;
-            row.addCell(SearchResultColumns.CWL_MAX_SIZE, resultRow[i] == null ? null : ((BigDecimal) resultRow[i]).toString());
+            row.addCell(SearchResultColumns.CWL_MAX_SIZE, resultRow[i] == null ? null : resultRow[i].toString());
             i++;
             row.addCell(SearchResultColumns.CWL_ID, (String) resultRow[i++]);
-            row.addCell(SearchResultColumns.AO_IDS_ACTUAL_COUNT, resultRow[i] == null ? null : ((BigDecimal) resultRow[i]).toString());
+            row.addCell(SearchResultColumns.AO_IDS_ACTUAL_COUNT, resultRow[i] == null ? null : resultRow[i].toString());
             i++;
-            row.addCell(SearchResultColumns.AO_WAITLIST_COUNT, resultRow[i] == null ? null : ((BigDecimal) resultRow[i]).toString());
+            row.addCell(SearchResultColumns.AO_WAITLIST_COUNT, resultRow[i] == null ? null : resultRow[i].toString());
 
             resultInfo.getRows().add(row);
         }
